Move PersistenceUnitUtil to the runtime module
